Move tests for serialization/deserialization in their own file
We didn't have a query-serialization.test.ts file. Now we do.

We can build on these tests when we begin to save other types of data.
