+++ extensions/ql-vscode/test/vscode-tests/no-workspace/query-results.test.ts
@@ -22,10 +22,6 @@ import { CodeQLCliServer, SourceInfo } from "../../../src/cli";
 import { CancellationTokenSource, Uri } from "vscode";
 import { tmpDir } from "../../../src/helpers";
 import {
-  deserializeQueryHistory,
-  serializeQueryHistory,
-} from "../../../src/query-serialization";
-import {
   formatLegacyMessage,
   QueryInProgress,
 } from "../../../src/legacy-query-server/run-queries";
@@ -436,117 +432,6 @@ describe("query-results", () => {
       },
       2 * 60 * 1000, // up to 2 minutes per test
     );
-  });
-
-  describe("serialize and deserialize", () => {
-    let infoSuccessRaw: LocalQueryInfo;
-    let infoSuccessInterpreted: LocalQueryInfo;
-    let infoEarlyFailure: LocalQueryInfo;
-    let infoLateFailure: LocalQueryInfo;
-    let infoInprogress: LocalQueryInfo;
-    let allHistory: LocalQueryInfo[];
-
-    beforeEach(() => {
-      infoSuccessRaw = createMockFullQueryInfo(
-        "a",
-        createMockQueryWithResults(
-          `${queryPath}-a`,
-          false,
-          false,
-          "/a/b/c/a",
-          false,
-        ),
-      );
-      infoSuccessInterpreted = createMockFullQueryInfo(
-        "b",
-        createMockQueryWithResults(
-          `${queryPath}-b`,
-          true,
-          true,
-          "/a/b/c/b",
-          false,
-        ),
-      );
-      infoEarlyFailure = createMockFullQueryInfo("c", undefined, true);
-      infoLateFailure = createMockFullQueryInfo(
-        "d",
-        createMockQueryWithResults(
-          `${queryPath}-c`,
-          false,
-          false,
-          "/a/b/c/d",
-          false,
-        ),
-      );
-      infoInprogress = createMockFullQueryInfo("e");
-      allHistory = [
-        infoSuccessRaw,
-        infoSuccessInterpreted,
-        infoEarlyFailure,
-        infoLateFailure,
-        infoInprogress,
-      ];
-    });
-
-    it("should serialize and deserialize query history", async () => {
-      // the expected results only contains the history with completed queries
-      const expectedHistory = [
-        infoSuccessRaw,
-        infoSuccessInterpreted,
-        infoLateFailure,
-      ];
-
-      const allHistoryPath = join(tmpDir.name, "workspace-query-history.json");
-
-      // serialize and deserialize
-      await serializeQueryHistory(allHistory, allHistoryPath);
-      const allHistoryActual = await deserializeQueryHistory(allHistoryPath);
-
-      // the dispose methods will be different. Ignore them.
-      allHistoryActual.forEach((info) => {
-        if (info.t === "local" && info.completedQuery) {
-          const completedQuery = info.completedQuery;
-          (completedQuery as any).dispose = undefined;
-
-          // these fields should be missing on the deserialized value
-          // but they are undefined on the original value
-          if (!("logFileLocation" in completedQuery)) {
-            (completedQuery as any).logFileLocation = undefined;
-          }
-          const query = completedQuery.query;
-          if (!("quickEvalPosition" in query)) {
-            (query as any).quickEvalPosition = undefined;
-          }
-        }
-      });
-      expectedHistory.forEach((info) => {
-        if (info.completedQuery) {
-          (info.completedQuery as any).dispose = undefined;
-        }
-      });
-
-      // make the diffs somewhat sane by comparing each element directly
-      for (let i = 0; i < allHistoryActual.length; i++) {
-        expect(allHistoryActual[i]).toEqual(expectedHistory[i]);
-      }
-      expect(allHistoryActual.length).toEqual(expectedHistory.length);
-    });
-
-    it("should handle an invalid query history version", async () => {
-      const badPath = join(tmpDir.name, "bad-query-history.json");
-      writeFileSync(
-        badPath,
-        JSON.stringify({
-          version: 3,
-          queries: allHistory,
-        }),
-        "utf8",
-      );
-
-      const allHistoryActual = await deserializeQueryHistory(badPath);
-      // version number is invalid. Should return an empty array.
-      expect(allHistoryActual).toEqual([]);
-    });
   });
 
   function safeDel(file: string) {
